Reworked the file system to provide different views based on the role of the user. Also reworked the nav bar to accomodate this update. Still some more navigation stuff to fix but it's progressing
